add some try-catch blocks
